CI: add 3.5 build with numpy_dev installed by wheel from master, #12057

+++ ci/install_conda.sh
@@ -81,6 +81,14 @@ conda info -a || exit 1
 # build deps
 REQ="ci/requirements-${TRAVIS_PYTHON_VERSION}${JOB_TAG}.build"
 time conda create -n pandas python=$TRAVIS_PYTHON_VERSION nose flake8 || exit 1
+
+# may have additional installation instructions for this build
+INSTALL="ci/install-${TRAVIS_PYTHON_VERSION}${JOB_TAG}.sh"
+if [ -e ${INSTALL} ]; then
+    time bash $INSTALL || exit 1
+fi
+
+# install deps
 time conda install -n pandas --file=${REQ} || exit 1
 
 source activate pandas
